+++ engine/apps/mobile_app/models.py
@@ -103,6 +103,7 @@ class MobileAppUserSettings(models.Model):
     important_notification_volume = models.FloatField(
         validators=[validators.MinValueValidator(0.0), validators.MaxValueValidator(1.0)], default=0.8
     )
+    important_notification_volume_override = models.BooleanField(default=True, null=True)
 
     # For the "Mobile push important" step it's possible to make notifications non-critical
     # if "override DND" setting is disabled in the app