+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
-### Added 
+### Added
 
 - Add mobile settings for info notifications by @imtoori ([#1926](https://github.com/grafana/oncall/pull/1926))